+++ src/cli.ts
@@ -1,14 +1,15 @@
 #!/usr/bin/env node
+import { loadConfig } from "graphile-config/load";
 import * as yargs from "yargs";
 
 import { defaults } from "./config";
 import getCronItems from "./getCronItems";
 import getTasks from "./getTasks";
 import { run, runOnce } from "./index";
-import { RunnerOptions } from "./interfaces";
+import { digestPreset } from "./lib";
 import { runMigrations } from "./runner";
 
-const argv_ = yargs
+const argv = yargs
   .parserConfiguration({
     "boolean-negation": false,
   })
@@ -70,28 +71,60 @@ const argv_ = yargs
     default: false,
   })
   .boolean("no-prepared-statements")
+  .option("config", {
+    alias: "C",
+    description: "The path to the config file",
+    normalize: true,
+  })
+  .string("config")
   .strict(true).argv;
 
-function isPromise(val: any): val is Promise<any> {
-  return typeof val === "object" && val && typeof val.then === "function";
-}
-
-// Hack TypeScript to stop whinging about argv potentially being a promise
-if (isPromise(argv_)) {
-  throw new Error("yargs returned a promise");
-}
-const argv = argv_ as Awaited<typeof argv_>;
-
-const isInteger = (n: number): boolean => {
-  return isFinite(n) && Math.round(n) === n;
+const isInteger = (n: number | undefined): boolean => {
+  return typeof n === "number" && isFinite(n) && Math.round(n) === n;
 };
 
+function stripUndefined<T extends object>(
+  t: T,
+): { [Key in keyof T as T[Key] extends undefined ? never : Key]: T[Key] } {
+  return Object.fromEntries(
+    Object.entries(t).filter(([_, value]) => value !== undefined),
+  ) as T;
+}
+
+function argvToPreset(inArgv: Awaited<typeof argv>): GraphileConfig.Preset {
+  return {
+    worker: stripUndefined({
+      connectionString: inArgv["connection"],
+      maxPoolSize: isInteger(inArgv["max-pool-size"])
+        ? inArgv["max-pool-size"]
+        : undefined,
+      pollInterval: isInteger(inArgv["poll-interval"])
+        ? inArgv["poll-interval"]
+        : undefined,
+      preparedStatements: !inArgv["no-prepared-statements"],
+      schema: inArgv.schema,
+      watch: inArgv.watch,
+      crontabFile: inArgv["crontab"],
+      concurrentJobs: isInteger(inArgv.jobs) ? inArgv.jobs : undefined,
+    }),
+  };
+}
+
 async function main() {
-  const DATABASE_URL = argv.connection || process.env.DATABASE_URL || undefined;
-  const SCHEMA = argv.schema || undefined;
+  const userPreset = await loadConfig(argv.config);
   const ONCE = argv.once;
   const SCHEMA_ONLY = argv["schema-only"];
-  const WATCH = argv.watch;
+  const {
+    runnerOptions: options,
+    resolvedPreset,
+    tasksFolder,
+    crontabFile,
+  } = digestPreset({
+    extends: [...(userPreset ? [userPreset] : []), argvToPreset(argv)],
+  });
+
+  const WATCH =
+    argv.watch || (!SCHEMA_ONLY && resolvedPreset.worker?.watch) || false;
 
   if (SCHEMA_ONLY && WATCH) {
     throw new Error("Cannot specify both --watch and --schema-only");
@@ -103,24 +136,11 @@ async function main() {
     throw new Error("Cannot specify both --watch and --once");
   }
 
-  if (!DATABASE_URL && !process.env.PGDATABASE) {
+  if (!options.connectionString) {
     throw new Error(
       "Please use `--connection` flag, set `DATABASE_URL` or `PGDATABASE` envvars to indicate the PostgreSQL connection to use.",
     );
   }
-
-  const options: RunnerOptions = {
-    schema: SCHEMA || defaults.schema,
-    concurrency: isInteger(argv.jobs) ? argv.jobs : defaults.concurrentJobs,
-    maxPoolSize: isInteger(argv["max-pool-size"])
-      ? argv["max-pool-size"]
-      : defaults.maxPoolSize,
-    pollInterval: isInteger(argv["poll-interval"])
-      ? argv["poll-interval"]
-      : defaults.pollInterval,
-    connectionString: DATABASE_URL,
-    noPreparedStatements: !!argv["no-prepared-statements"],
-  };
 
   if (SCHEMA_ONLY) {
     await runMigrations(options);
@@ -128,12 +148,8 @@ async function main() {
     return;
   }
 
-  const watchedTasks = await getTasks(options, `${process.cwd()}/tasks`, WATCH);
-  const watchedCronItems = await getCronItems(
-    options,
-    argv.crontab || `${process.cwd()}/crontab`,
-    WATCH,
-  );
+  const watchedTasks = await getTasks(options, tasksFolder, WATCH);
+  const watchedCronItems = await getCronItems(options, crontabFile, WATCH);
 
   if (ONCE) {
     await runOnce(options, watchedTasks.tasks);