+++ src/cron.ts
@@ -110,6 +110,8 @@ async function scheduleCronJobs(
   ts: string,
   useNodeTime: boolean,
 ) {
+  // TODO: refactor this to use `add_jobs`
+
   // Note that `identifier` is guaranteed to be unique for every record
   // in `specs`.
   await pgPool.query(
@@ -132,8 +134,8 @@ async function scheduleCronJobs(
         insert into ${escapedWorkerSchema}.known_crontabs (identifier, known_since, last_execution)
         select
           specs.identifier,
-          $2 as known_since,
-          $2 as last_execution
+          $2::timestamptz as known_since,
+          $2::timestamptz as last_execution
         from specs
         on conflict (identifier)
         do update set last_execution = excluded.last_execution
@@ -188,7 +190,7 @@ async function registerAndBackfillItems(
     await pgPool.query(
       `
       INSERT INTO ${escapedWorkerSchema}.known_crontabs (identifier, known_since)
-      SELECT identifier, $2
+      SELECT identifier, $2::timestamptz
       FROM unnest($1::text[]) AS unnest (identifier)
       ON CONFLICT DO NOTHING
       `,
@@ -293,7 +295,7 @@ export const runCron = (
 
   const promise = defer();
   let released = false;
-  let timeout: NodeJS.Timer | null = null;
+  let timeout: NodeJS.Timeout | null = null;
 
   let stopCalled = false;
   function stop(e?: Error) {
@@ -342,7 +344,8 @@ export const runCron = (
     // The backfill may have taken a moment, we should continue from where the
     // worker started and catch up as quickly as we can. This does **NOT**
     // count as a backfill.
-    let nextTimestamp = unsafeRoundToMinute(new Date(+start), true);
+    /** This timestamp will be mutated! */
+    const nextTimestamp = unsafeRoundToMinute(new Date(+start), true);
 
     const scheduleNextLoop = () => {
       if (released) {
@@ -497,18 +500,18 @@ export async function getParsedCronItemsFromOptions(
   }
 
   if (crontab) {
-    assert(
+    assert.ok(
       !crontabFile,
       "`crontab` and `crontabFile` must not be set at the same time.",
     );
-    assert(
+    assert.ok(
       !parsedCronItems,
       "`crontab` and `parsedCronItems` must not be set at the same time.",
     );
 
     return parseCrontab(crontab);
   } else if (crontabFile) {
-    assert(
+    assert.ok(
       !parsedCronItems,
       "`crontabFile` and `parsedCronItems` must not be set at the same time.",
     );
@@ -517,14 +520,14 @@ export async function getParsedCronItemsFromOptions(
     releasers.push(() => watchedCronItems.release());
     return watchedCronItems.items;
   } else {
-    assert(parsedCronItems != null, "Expected `parsedCronItems` to be set.");
+    assert.ok(parsedCronItems != null, "Expected `parsedCronItems` to be set.");
     // Basic check to ensure that users remembered to call
     // `parseCronItems`/`parseCrontab`; not intended to be a full check, just a
     // quick one to catch the obvious errors. Keep in mind that
     // `parsedCronItems` is mutable so it may be changed later to contain more
     // entries; we can't keep performing these checks everywhere for
     // performance reasons.
-    assert(
+    assert.ok(
       Array.isArray(parsedCronItems),
       "Expected `parsedCronItems` to be an array; you must use a helper e.g. `parseCrontab()` or `parseCronItems()` to produce this value.",
     );