+++ src/helpers.ts
@@ -25,9 +25,9 @@ export function makeAddJob(
           payload => $2::json,
           queue_name => $3::text,
           run_at => $4::timestamptz,
-          max_attempts => $5::smallint,
+          max_attempts => $5::int,
           job_key => $6::text,
-          priority => $7::smallint,
+          priority => $7::int,
           flags => $8::text[],
           job_key_mode => $9::text
         );
@@ -85,13 +85,13 @@ export function makeJobHelpers(
 
   // DEPRECATED METHODS
   Object.assign(helpers, {
-    debug(format: string, ...parameters: any[]): void {
+    debug(format: string, ...parameters: unknown[]): void {
       logger.error(
         "REMOVED: `helpers.debug` has been replaced with `helpers.logger.debug`; please do not use `helpers.debug`",
       );
       logger.debug(format, { parameters });
     },
-  } as any);
+  } as unknown);
 
   return helpers;
 }