+++ src/main.ts
@@ -19,7 +19,7 @@ import {
 } from "./interfaces";
 import { processSharedOptions } from "./lib";
 import { Logger } from "./logger";
-import SIGNALS from "./signals";
+import SIGNALS, { Signal } from "./signals";
 import { failJobs } from "./sql/failJob";
 import { resetLockedAt } from "./sql/resetLockedAt";
 import { makeNewWorker } from "./worker";
@@ -40,7 +40,7 @@ export { allWorkerPools as _allWorkerPools };
  * gracefulShutdown to all the pools' events; we use this event emitter to
  * aggregate these requests.
  */
-let _signalHandlersEventEmitter: WorkerEvents = new EventEmitter();
+const _signalHandlersEventEmitter: WorkerEvents = new EventEmitter();
 
 /**
  * Only register the signal handlers once _globally_.
@@ -50,7 +50,8 @@ let _registeredSignalHandlers = false;
 /**
  * Only trigger graceful shutdown once.
  */
-let _shuttingDown = false;
+let _shuttingDownGracefully = false;
+let _shuttingDownForcefully = false;
 
 /**
  * This will register the signal handlers to make sure the worker shuts down
@@ -59,48 +60,116 @@ let _shuttingDown = false;
  * future calls will register the events but take no further actions.
  */
 function registerSignalHandlers(logger: Logger, events: WorkerEvents) {
-  if (_shuttingDown) {
+  if (_shuttingDownGracefully || _shuttingDownForcefully) {
     throw new Error(
       "System has already gone into shutdown, should not be spawning new workers now!",
     );
   }
+
   _signalHandlersEventEmitter.on("gracefulShutdown", (o) =>
     events.emit("gracefulShutdown", o),
   );
+  _signalHandlersEventEmitter.on("forcefulShutdown", (o) =>
+    events.emit("forcefulShutdown", o),
+  );
+
   if (_registeredSignalHandlers) {
     return;
+  } else {
+    _registeredSignalHandlers = true;
   }
-  _registeredSignalHandlers = true;
-  SIGNALS.forEach((signal) => {
-    logger.debug(`Registering signal handler for ${signal}`, {
-      registeringSignalHandler: signal,
-    });
-    const removeHandler = () => {
-      logger.debug(`Removing signal handler for ${signal}`, {
-        unregisteringSignalHandler: signal,
-      });
-      process.removeListener(signal, handler);
-    };
-    const handler = function () {
-      logger.error(`Received '${signal}'; attempting graceful shutdown...`);
-      setTimeout(removeHandler, 5000);
-      if (_shuttingDown) {
-        return;
-      }
-      _shuttingDown = true;
-      _signalHandlersEventEmitter.emit("gracefulShutdown", { signal });
-      Promise.all(
-        allWorkerPools.map((pool) =>
-          pool.gracefulShutdown(`Forced worker shutdown due to ${signal}`),
-        ),
-      ).finally(() => {
-        removeHandler();
-        logger.error(`Graceful shutdown attempted; killing self via ${signal}`);
+  const switchToForcefulHandler = () => {
+    logger.debug(
+      `Switching to forceful handler for termination signals (${SIGNALS.join(
+        ", ",
+      )}); another termination signal will force a fast (unsafe) shutdown`,
+      { switchToForcefulHandlers: true },
+    );
+    for (const signal of SIGNALS) {
+      process.on(signal, forcefulHandler);
+      process.removeListener(signal, gracefulHandler);
+    }
+  };
+  const removeForcefulHandler = () => {
+    logger.debug(
+      `Removed forceful handler for termination signals (${SIGNALS.join(
+        ", ",
+      )}); another termination signals will likely kill the process (unless you've registered other handlers)`,
+      { unregisteringSignalHandlers: true },
+    );
+    for (const signal of SIGNALS) {
+      process.removeListener(signal, forcefulHandler);
+    }
+  };
+
+  const gracefulHandler = function (signal: Signal) {
+    if (_shuttingDownGracefully) {
+      logger.error(
+        `Ignoring '${signal}' (graceful shutdown already in progress)`,
+      );
+      return;
+    } else {
+      _shuttingDownGracefully = true;
+    }
+
+    logger.error(
+      `Received '${signal}'; attempting global graceful shutdown... (all termination signals will be ignored for the next 5 seconds)`,
+    );
+    const switchTimeout = setTimeout(switchToForcefulHandler, 5000);
+    _signalHandlersEventEmitter.emit("gracefulShutdown", { signal });
+
+    Promise.allSettled(
+      allWorkerPools.map((pool) =>
+        pool.gracefulShutdown(`Graceful worker shutdown due to ${signal}`),
+      ),
+    ).finally(() => {
+      clearTimeout(switchTimeout);
+      process.removeListener(signal, gracefulHandler);
+      if (!_shuttingDownForcefully) {
+        logger.error(
+          `Global graceful shutdown complete; killing self via ${signal}`,
+        );
         process.kill(process.pid, signal);
-      });
-    };
-    process.on(signal, handler);
-  });
+      }
+    });
+  };
+  const forcefulHandler = function (signal: Signal) {
+    if (_shuttingDownForcefully) {
+      logger.error(
+        `Ignoring '${signal}' (forceful shutdown already in progress)`,
+      );
+      return;
+    } else {
+      _shuttingDownForcefully = true;
+    }
+
+    logger.error(
+      `Received '${signal}'; attempting global forceful shutdown... (all termination signals will be ignored for the next 5 seconds)`,
+    );
+    const removeTimeout = setTimeout(removeForcefulHandler, 5000);
+    _signalHandlersEventEmitter.emit("forcefulShutdown", { signal });
+
+    Promise.allSettled(
+      allWorkerPools.map((pool) =>
+        pool.forcefulShutdown(`Forced worker shutdown due to ${signal}`),
+      ),
+    ).finally(() => {
+      removeForcefulHandler();
+      clearTimeout(removeTimeout);
+      logger.error(
+        `Global forceful shutdown completed; killing self via ${signal}`,
+      );
+      process.kill(process.pid, signal);
+    });
+  };
+
+  logger.debug(
+    `Registering termination signal handlers (${SIGNALS.join(", ")})`,
+    { registeringSignalHandlers: SIGNALS },
+  );
+  for (const signal of SIGNALS) {
+    process.on(signal, gracefulHandler);
+  }
 }
 
 export function runTaskList(
@@ -122,23 +191,25 @@ export function runTaskList(
   const promise = deferred();
   const workers: Array<Worker> = [];
 
-  let listenForChangesClient: PoolClient | null = null;
+  let changeListener: {
+    client: PoolClient;
+    release: () => void;
+  } | null = null;
 
   const unlistenForChanges = async () => {
-    if (listenForChangesClient) {
-      const client = listenForChangesClient;
-      listenForChangesClient = null;
-      // Unsubscribe from jobs:insert topic
+    if (changeListener) {
       try {
-        await client.query('UNLISTEN "jobs:insert"');
+        changeListener.release();
       } catch (e) {
-        // Ignore
+        logger.error(
+          `Error occurred whilst releasing listening client: ${e.message}`,
+          { error: e },
+        );
       }
-      await client.release();
     }
   };
   let active = true;
-  let reconnectTimeout: NodeJS.Timer | null = null;
+  let reconnectTimeout: NodeJS.Timeout | null = null;
 
   const compiledSharedOptions = processSharedOptions(options);
   const { minResetLockedInterval, maxResetLockedInterval } =
@@ -205,63 +276,174 @@ export function runTaskList(
     Math.random() * Math.min(60000, maxResetLockedInterval),
   );
 
-  // This is a representation of us that can be interacted with externally
-  const workerPool: WorkerPool = {
-    release: async () => {
-      // IMPORTANT: if we assert that `active === true` here, we must ensure this is handled in `gracefulShutdown`
+  function deactivate() {
+    if (active) {
       active = false;
-      clearTimeout(resetLockedTimeout!);
-      resetLockedTimeout = null;
+      if (resetLockedTimeout) {
+        clearTimeout(resetLockedTimeout);
+        resetLockedTimeout = null;
+      }
       if (reconnectTimeout) {
         clearTimeout(reconnectTimeout);
         reconnectTimeout = null;
       }
       events.emit("pool:release", { pool: this });
       unlistenForChanges();
-      promise.resolve(resetLockedAtPromise);
-      await Promise.all(workers.map((worker) => worker.release()));
+    }
+  }
+
+  let terminated = false;
+  function terminate() {
+    if (!terminated) {
+      terminated = true;
       const idx = allWorkerPools.indexOf(workerPool);
       allWorkerPools.splice(idx, 1);
+      promise.resolve(resetLockedAtPromise);
+    } else {
+      logger.error(
+        `Graphile Worker internal error: terminate() was called twice for worker pool. Ignoring second call; but this indicates a bug - please file an issue.`,
+      );
+    }
+  }
+
+  // This is a representation of us that can be interacted with externally
+  const workerPool: WorkerPool = {
+    release: async () => {
+      console.trace(
+        "DEPRECATED: You are calling `workerPool.release()`; please use `workerPool.gracefulShutdown()` instead.",
+      );
+      return this.gracefulShutdown();
     },
 
-    // Make sure we clean up after ourselves even if a signal is caught
-    async gracefulShutdown(message: string) {
+    /**
+     * Stop accepting jobs, and wait gracefully for the jobs that are in
+     * progress to complete.
+     */
+    async gracefulShutdown(
+      message = "Worker pool is shutting down gracefully",
+    ) {
       events.emit("pool:gracefulShutdown", { pool: this, message });
       try {
         logger.debug(`Attempting graceful shutdown`);
         // Stop new jobs being added
-        active = false;
+        deactivate();
 
-        // Release all our workers' jobs
-        const workerIds = workers.map((worker) => worker.workerId);
-        const jobsInProgress: Array<Job> = workers
-          .map((worker) => worker.getActiveJob())
-          .filter((job): job is Job => !!job);
         // Remove all the workers - we're shutting them down manually
-        workers.splice(0, workers.length).map((worker) => worker.release());
-        logger.debug(`Releasing the jobs '${workerIds.join(", ")}'`, {
-          workerIds,
-        });
-        const cancelledJobs = await failJobs(
-          compiledSharedOptions,
-          withPgClient,
-          workerIds,
-          jobsInProgress,
-          message,
-        );
-        logger.debug(`Cancelled ${cancelledJobs.length} jobs`, {
-          cancelledJobs,
-        });
-        logger.debug("Jobs released");
+        const workerPromises = workers.map((worker) => worker.release());
+        const workerReleaseResults = await Promise.allSettled(workerPromises);
+        const jobsToRelease: Job[] = [];
+        for (let i = 0; i < workerReleaseResults.length; i++) {
+          const workerReleaseResult = workerReleaseResults[i];
+          if (workerReleaseResult.status === "rejected") {
+            const worker = workers[i];
+            const job = worker.getActiveJob();
+            events.emit("pool:gracefulShutdown:workerError", {
+              pool: this,
+              error: workerReleaseResult.reason,
+              job,
+            });
+            logger.debug(
+              `Cancelling worker ${worker.workerId} (job: ${
+                job?.id ?? "none"
+              }) failed`,
+              {
+                worker,
+                job,
+                reason: workerReleaseResult.reason,
+              },
+            );
+            if (job) {
+              jobsToRelease.push(job);
+            }
+          }
+        }
+        if (jobsToRelease.length > 0) {
+          const workerIds = workers.map((worker) => worker.workerId);
+          logger.debug(
+            `Releasing the jobs ${jobsToRelease
+              .map((j) => j.id)
+              .join()} (workers: ${workerIds.join(", ")})`,
+            {
+              jobs: jobsToRelease,
+              workerIds,
+            },
+          );
+          const cancelledJobs = await failJobs(
+            compiledSharedOptions,
+            withPgClient,
+            workerIds,
+            jobsToRelease,
+            message,
+          );
+          logger.debug(`Cancelled ${cancelledJobs.length} jobs`, {
+            cancelledJobs,
+          });
+        }
+        events.emit("pool:gracefulShutdown:complete", { pool: this });
+        logger.debug("Graceful shutdown complete");
       } catch (e) {
         events.emit("pool:gracefulShutdown:error", { pool: this, error: e });
         logger.error(`Error occurred during graceful shutdown: ${e.message}`, {
           error: e,
         });
+        return this.forcefulShutdown(e.message);
       }
+      terminate();
+    },
 
-      // Remove ourself from the list of worker pools
-      await this.release();
+    /**
+     * Stop accepting jobs and "fail" all currently running jobs.
+     */
+    async forcefulShutdown(message: string) {
+      events.emit("pool:forcefulShutdown", { pool: this, message });
+      try {
+        logger.debug(`Attempting forceful shutdown`);
+        // Stop new jobs being added
+        deactivate();
+
+        // Release all our workers' jobs
+        const jobsInProgress: Array<Job> = workers
+          .map((worker) => worker.getActiveJob())
+          .filter((job): job is Job => !!job);
+
+        // Remove all the workers - we're shutting them down manually
+        const workerPromises = workers.map((worker) => worker.release());
+        // Ignore the results, we're shutting down anyway
+        Promise.allSettled(workerPromises);
+
+        if (jobsInProgress.length > 0) {
+          const workerIds = workers.map((worker) => worker.workerId);
+          logger.debug(
+            `Releasing the jobs ${jobsInProgress
+              .map((j) => j.id)
+              .join()} (workers: ${workerIds.join(", ")})`,
+            {
+              jobs: jobsInProgress,
+              workerIds,
+            },
+          );
+          const cancelledJobs = await failJobs(
+            compiledSharedOptions,
+            withPgClient,
+            workerIds,
+            jobsInProgress,
+            message,
+          );
+          logger.debug(`Cancelled ${cancelledJobs.length} jobs`, {
+            cancelledJobs,
+          });
+        } else {
+          logger.debug("No active jobs to release");
+        }
+        events.emit("pool:forcefulShutdown:complete", { pool: this });
+        logger.debug("Forceful shutdown complete");
+      } catch (e) {
+        events.emit("pool:forcefulShutdown:error", { pool: this, error: e });
+        logger.error(`Error occurred during forceful shutdown: ${e.message}`, {
+          error: e,
+        });
+      }
+      terminate();
     },
 
     promise,
@@ -325,7 +507,6 @@ export function runTaskList(
         return;
       }
       errorHandled = true;
-      listenForChangesClient = null;
       try {
         release();
       } catch (e) {
@@ -338,13 +519,14 @@ export function runTaskList(
     }
 
     function handleNotification() {
-      if (listenForChangesClient === client) {
+      if (changeListener?.client === client) {
         // Find a worker that's available
         workers.some((worker) => worker.nudge());
       }
     }
 
     function release() {
+      changeListener = null;
       client.removeListener("error", onErrorReleaseClientAndTryAgain);
       client.removeListener("notification", handleNotification);
       client.query('UNLISTEN "jobs:insert"').catch(() => {
@@ -359,7 +541,7 @@ export function runTaskList(
     //----------------------------------------
 
     events.emit("pool:listen:success", { workerPool, client });
-    listenForChangesClient = client;
+    changeListener = { client, release };
     client.on("notification", handleNotification);
 
     // Subscribe to jobs:insert message
@@ -407,8 +589,8 @@ export const runTaskListOnce = (
       makeWithPgClientFromClient(client),
       false,
     );
-    finalPromise["worker"] = worker;
+    finalPromise.worker = worker;
     return worker.promise;
-  });
+  }) as Promise<void> & { worker: Worker };
   return finalPromise;
 };