+++ src/sql/resetLockedAt.ts
@@ -19,7 +19,7 @@ export async function resetLockedAt(
       text: `\
 with j as (
 update ${escapedWorkerSchema}.jobs
-set locked_at = null, locked_by = null
+set locked_at = null, locked_by = null, run_at = greatest(run_at, now())
 where locked_at < ${now} - interval '4 hours'
 )
 update ${escapedWorkerSchema}.job_queues
@@ -28,7 +28,7 @@ where locked_at < ${now} - interval '4 hours'`,
       values: useNodeTime ? [new Date().toISOString()] : [],
       name: noPreparedStatements
         ? undefined
-        : `clear_stale_locks/${workerSchema}`,
+        : `clear_stale_locks${useNodeTime ? "N" : ""}/${workerSchema}`,
     }),
   );
 }