+++ website/docusaurus.config.js
@@ -1,0 +1,122 @@
+// @ts-check
+
+/* eslint-disable @typescript-eslint/no-var-requires */
+
+const lightCodeTheme = require("prism-react-renderer").themes.github;
+const darkCodeTheme = require("prism-react-renderer").themes.dracula;
+
+/** @type {import('@docusaurus/types').Config} */
+const config = {
+  title: "Graphile Worker",
+  tagline: "A high performance job queue for PostgreSQL, written in Node.js",
+  /*
+  <link rel="apple-touch-icon" sizes="180x180" href="/img/apple-touch-icon.png">
+  <link rel="icon" type="image/png" sizes="32x32" href="/img/favicon-32x32.png">
+  <link rel="icon" type="image/png" sizes="16x16" href="/img/favicon-16x16.png">
+  <link rel="manifest" href="/img/site.webmanifest">
+  <link rel="mask-icon" href="/img/safari-pinned-tab.svg" color="#5bbad5">
+  <link rel="shortcut icon" href="/img/favicon.ico">
+  <meta name="msapplication-TileColor" content="#da532c">
+  <meta name="msapplication-config" content="/img/browserconfig.xml">
+  <meta name="theme-color" content="#ffffff">
+  */
+  favicon: "img/favicon.ico",
+
+  url: "https://worker.graphile.org",
+  baseUrl: "/",
+
+  organizationName: "graphile",
+  projectName: "worker",
+  deploymentBranch: "gh-pages",
+  trailingSlash: false,
+
+  onBrokenLinks: "throw",
+  onBrokenMarkdownLinks: "warn",
+
+  i18n: {
+    defaultLocale: "en",
+    locales: ["en"],
+  },
+
+  presets: [
+    [
+      "classic",
+      /** @type {import('@docusaurus/preset-classic').Options} */
+      ({
+        docs: {
+          sidebarPath: require.resolve("./sidebars.js"),
+          editUrl: "https://github.com/graphile/worker/tree/main/website/",
+          remarkPlugins: [
+            [require("@docusaurus/remark-plugin-npm2yarn"), { sync: true }],
+          ],
+        },
+        pages: {
+          remarkPlugins: [
+            [require("@docusaurus/remark-plugin-npm2yarn"), { sync: true }],
+          ],
+        },
+        blog: false /*{
+          showReadingTime: true,
+          editUrl: "https://github.com/graphile/worker/tree/main/website/",
+        }*/,
+        theme: {
+          customCss: require.resolve("./src/css/custom.css"),
+        },
+      }),
+    ],
+  ],
+
+  themeConfig:
+    /** @type {import('@docusaurus/preset-classic').ThemeConfig} */
+    ({
+      // image: "img/docusaurus-social-card.jpg",
+      navbar: {
+        title: "Graphile Worker",
+        logo: {
+          alt: "Graphile Worker",
+          src: "img/logo.svg",
+        },
+        items: [
+          {
+            href: "https://github.com/graphile/worker",
+            label: "GitHub",
+            position: "right",
+          },
+        ],
+      },
+      footer: {
+        style: "dark",
+        links: [
+          {
+            title: "Community",
+            items: [
+              {
+                label: "Discord",
+                href: "https://discord.gg/graphile",
+              },
+              {
+                label: "Twitter",
+                href: "https://twitter.com/GraphileHQ",
+              },
+            ],
+          },
+          {
+            title: "More",
+            items: [
+              {
+                label: "GitHub",
+                href: "https://github.com/graphile/worker",
+              },
+            ],
+          },
+        ],
+        copyright: `Copyright © ${new Date().getFullYear()} Graphile Ltd. Built with Docusaurus.`,
+      },
+      prism: {
+        theme: lightCodeTheme,
+        darkTheme: darkCodeTheme,
+      },
+    }),
+};
+
+module.exports = config;